mcp-servers/tracker/requirements.txt
fastmcp>=2.0.0
gspread>=6.0.0
google-auth>=2.0.0
